integration: move batch LWT tests to batch.rs
